requirements.txt
requests==2.21.0
cloudinary==1.18.2
uperations_kernel==0.0.10
uperations_base==0.0.20
mysql_connector_repackaged==0.3.1
